requirements_colab.txt
# =============================================================================
# requirements_colab.txt
#
# วิธีใช้ใน Google Colab (ต้น notebook):
#   !pip install -q -r requirements_colab.txt
#
# หมายเหตุ:
#   - PyTorch / torchvision ไม่อยู่ในไฟล์นี้ เพราะ Colab มี CUDA version
#     ให้อยู่แล้ว และการติดตั้งซ้ำทำให้ version ชนได้
#   - Package ที่เป็น Windows-only (pywin32, colorama) ถูกตัดออก
#   - Package ที่ Colab มีให้แล้ว (numpy, pillow, matplotlib ฯลฯ)
#     ระบุ version เพื่อ pin ให้ตรง
# =============================================================================

# ── Core ML / Vision ──────────────────────────────────────────────────────────
anomalib==2.2.0
timm==1.0.25
kornia==0.8.2
kornia_rs==0.1.10
einops==0.8.2
open3d==0.19.0

# ── Scientific Stack ──────────────────────────────────────────────────────────
numpy==2.2.6
scipy==1.15.3
scikit-learn==1.7.2
scikit-image==0.25.2
pandas==2.3.3

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python==4.13.0.92
imagecodecs==2025.3.30
ImageIO==2.37.2
tifffile==2025.5.10
Pillow==12.1.1

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib==3.10.8
seaborn==0.13.2
plotly==6.5.2

# ── Lightning / Training Utilities ────────────────────────────────────────────
lightning==2.6.1
pytorch-lightning==2.6.1
lightning-utilities==0.15.3
torchmetrics==1.8.2

# ── Config / CLI ──────────────────────────────────────────────────────────────
omegaconf==2.3.0
jsonargparse==4.46.0
ConfigArgParse==1.7.1
PyYAML==6.0.3
typer==0.24.1
rich==14.3.3
rich-argparse==1.7.2

# ── Normalising Flows (for anomalib) ──────────────────────────────────────────
FrEIA==0.2

# ── Mesh / Geometry ───────────────────────────────────────────────────────────
robust_laplacian==1.0.0

# ── Web / Dashboard ───────────────────────────────────────────────────────────
dash==4.0.0
Flask==3.1.3

# ── Hugging Face ──────────────────────────────────────────────────────────────
huggingface_hub==1.5.0
safetensors==0.7.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm==4.67.3
joblib==1.5.3
networkx==3.4.2
sympy==1.14.0
mpmath==1.3.0
requests==2.32.5
psutil==7.2.2
